requirements.txt
flask
requests
stripe
airtable-python-wrapper